pyproject.toml
[project]
name = "genshin-map"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.3.0",
    "requests>=2.32.5",
    "rich>=14.1.0",
    "typer>=0.17.4",
]
